# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Mini financial transfer API with atomicity, idempotency, and batch processing.",
        "title": "Cita Kita Transfer Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
        "/accounts": {
            "get": {
                "description": "Returns all seeded users with their current balances",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "List accounts",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.AccountListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/batch-transfer": {
            "post": {
                "description": "Processes multiple transfers concurrently. Each item is idempotent by idempotency_key.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transfer"
                ],
                "summary": "Batch transfer funds",
                "parameters": [
                    {
                        "description": "Batch transfer requests",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.TransferRequest"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.BatchTransferResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Returns OK when the service is running",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/transfer": {
            "post": {
                "description": "Transfers amount from one account to another. Idempotent by idempotency_key.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transfer"
                ],
                "summary": "Transfer funds between accounts",
                "parameters": [
                    {
                        "description": "Transfer request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.TransferRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.TransferResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.AccountListResponse": {
            "type": "object",
            "properties": {
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.User"
                    }
                }
            }
        },
        "model.BatchDetail": {
            "type": "object",
            "properties": {
                "idempotency_key": {
                    "type": "string",
                    "example": "p3Rt8yKm2QnX5vJ9"
                },
                "message": {
                    "type": "string",
                    "example": "transfer OK"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "success",
                        "failed"
                    ],
                    "example": "success"
                }
            }
        },
        "model.BatchTransferResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.BatchDetail"
                    }
                },
                "failed": {
                    "type": "integer",
                    "example": 1
                },
                "success": {
                    "type": "integer",
                    "example": 2
                }
            }
        },
        "model.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid request"
                }
            }
        },
        "model.TransferRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer",
                    "example": 100
                },
                "from_id": {
                    "type": "string",
                    "example": "A1"
                },
                "idempotency_key": {
                    "type": "string",
                    "example": "k7Fp2mQx9ZnW4bL8"
                },
                "to_id": {
                    "type": "string",
                    "example": "A2"
                }
            }
        },
        "model.TransferResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer",
                    "example": 100
                },
                "from_id": {
                    "type": "string",
                    "example": "A1"
                },
                "idempotency_key": {
                    "type": "string",
                    "example": "k7Fp2mQx9ZnW4bL8"
                },
                "to_id": {
                    "type": "string",
                    "example": "A2"
                }
            }
        },
        "model.User": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "integer",
                    "example": 1000
                },
                "id": {
                    "type": "string",
                    "example": "A1"
                },
                "name": {
                    "type": "string",
                    "example": "John Doe"
                }
            }
        }
    }
}